+++ axum/src/extract/mod.rs
@@ -11,6 +11,7 @@ pub mod rejection;
 pub mod ws;
 
 mod host;
+mod raw_form;
 mod raw_query;
 mod request_parts;
 mod state;
@@ -19,13 +20,14 @@ mod state;
 pub use axum_core::extract::{DefaultBodyLimit, FromRef, FromRequest, FromRequestParts};
 
 #[cfg(feature = "macros")]
-pub use axum_macros::{FromRequest, FromRequestParts};
+pub use axum_macros::{FromRef, FromRequest, FromRequestParts};
 
 #[doc(inline)]
 #[allow(deprecated)]
 pub use self::{
     host::Host,
     path::Path,
+    raw_form::RawForm,
     raw_query::RawQuery,
     request_parts::{BodyStream, RawBody},
     state::State,